refactored locales, forms and interfaces
